scranton/agents/dwight_schrute/requirements.txt
google-adk>=1.3.0
python-dotenv>=1.1.0
litellm>=1.72.7
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.41